Update a_star_levi_julia.py
These additions were focused on checking the open loop and closed loop for repeats.  I used the second method from the instructions with a two matrices the size of the workspace times two plus a section for the angle.  I then check_cl and check_ol are used for the if statements and if either is zero they are changed to 1 after the if statement to show they have been explored.