Update a_star_levi_julia.py
I added Julia's code for defining the obstacle space.  It uses two for loops to create an a matrix where 0 is a free space and 1 is in the obstacle space. It has been double to account for the 0.5mm resolution we are using.  That will have to be changed later