Update a_star_levi_julia.py
I added Julia's function for taking in user input and checking if it is in the obstacle space and if the input step size of the robot is usable.  Also I added her line of code that calculates the initial cost to go, which is the linear distance from the start to the goal.