Update a_star_levi_julia.py
As per Julia, decoupled checking the obstacle as we can skip updating code if the new node is in the obstacle space.  Modified the updated cost with her updated calculations.  The cost to go makes it such that you cannot simply add to the current cost, as cost to come is dependent on the position.