Update a_star_levi_julia.py
After the goal node is reached, the path needs to be traced back to the parent node.  I accomplished this with two functions.  One function (trace_back) bounces back and forth finding a parent and corresponding index until the parent node is found.  Afterwards, the function reverse list is used to order the list from start node to final node.