Update a_star_levi_julia.py
This is an update to how a repeat in the priority queue is updated.  If the index of a priority queue is changed, it no longer functions as a priority queue.  To preserve the correct order in a priority queue the repeat node must be removed and then the improved node must put in the priority queue.