Update a_star_levi_julia.py
I added a return to the traceback function as that was missing.  Based on code from my djikstra algorithm.  I pulled the x and y coordinates from the closed list, open list, and optimal path.  The values with an extra th after the x or y are used for the quiver function.  They give the orientation and direction of the arrows.  In a for loop, I did a test plot of the closed loop arrows with a scale of 1.  I may change the scale of the arrows to ensure the end of one arrow touches the end of the next arrow.