Update a_star_levi_julia.py
I updated how the cost was to come was calculated.  I simplified it to the cost to come of the previous node plus the step size of the robot.  I also changed to code for how we find the index of a repeated node in the open loop.  As much as I would like to avoid a for loop, that's how I knew how to fix it.  It still took under 1 minute to get from one side of the map to the other.  Need to fix the animation, though.