README clarification and start of new file
I added a sentence to the README file to clarify where the video for the project is.  I wanted to save the grader time while grading.  Also, I made a new file for project 3 phase 2.  I added some of the packages that need to be added and some useful values from the spec sheet.