Update proj3p2_levi_julia.py
I added in the inputs for this program based on what Julia wrote for A* and the prompt for project 3 phase 2.  It takes in the clearance, start state, goal state, and wheel rpms.  I put the clearance first as I will need that to calculate the obstacle space , and that is need for the start and goal nodes.