Update proj3p2_levi_julia.py
I added the action set vector (the rpm options.  I moved the variables from the turtlebot data sheet to before the functions.  I wrote a function that takes in the velocity from the left wheel, the velocity from the right wheel, the dt value, and the theta value, and returns the change in x, change in y, and change in theta.