Update proj3p2_levi_julia.py
I copied over and modified all of the code for plotting the final result from the A* code we hade written.  It utilizes quiver from Matplotlib.  I also fixed the bounds and the graph makes it easy to spot mistakes.  The closed loop is not plotting for some reason, so I will have to fix that later.