Update proj3p2_levi_julia.py
I fixed the remaining issues in finding the obstacle space and graphing the obstacle space.  I increased the dt vale for 0.2 to 1. I also increased the number of points displayed per frame.  This results in a code that works okay, but not perfectly.  I don't know how I feel about the visuals