Modifications for 3D Visualization
I changed the dt of the main python file to 1 second.  I also removed the part of the code that aborts the program if it takes too long.  It has been running quite fast.  I created a new file that I am working on creating a new program based on the main file.  This records the x and y velocity over time.  I will added a talker function later.